% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{player_totals}
\alias{player_totals}
\title{Season summary statistics of all NBA players from the 2006-2007 season to the
2019-2020 season.}
\format{
A data frame with 6715 rows and 40 variables
}
\source{
Scraped from \url{https://www.basketball-reference.com/}.
}
\usage{
player_totals
}
\description{
A dataset that is directly scraped from Basketball Reference and presented
unprocessed except for one additional column, GroupPosition:
most frequently played group position during study period according to
Basketball Reference (G, F or C)
}
\keyword{datasets}
